Update patch set 1
Patch Set 1: Code-Review+2

Max, what I can't figure out is why we haven't gotten reports of it against the n1.

My guess is that it might depend on the SIM inserted, specifically on the data APN, since it appears related to SUPL AGPS.

I'll try adding some logging when I test this on the find5 tonight or tomorrow.

Patch-set: 1
Label: Code-Review=+2
Label: SUBM=+1